+++ Properties/AssemblyInfo.cs
@@ -15,6 +15,7 @@ using System.Runtime.InteropServices;
                  "\nHighlights are methods such as:" +
                  "\nUse.Save() and.Load() for easy state saving of any object in C#." +
                  "\nUse.Retry() to easily handle 429 throttling errors in web calls." +
+                 "\nUse System.Timer class to quickly add stopwatch timing functionality." +
                  "\nUse.IsPrime() to determine if a number is a prime number." +
                  "\nUse.GetNthPrime() and.GetNthPrimeAsync() to get the Nth prime number capable of calculating 10, 000, 000 primes / minute on an Intel Core i7 - 6700K CPU @ 4.00 GHz using 7 threads on 8 logical cores." +
                  "\nUse.ReplaceTokens() to replace multiple strings in string dictionary style." +
@@ -48,7 +49,7 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.1")]
-[assembly: AssemblyVersion("3.1.505")]
-[assembly: AssemblyFileVersion("3.1.505")]
+[assembly: AssemblyVersion("3.2.500")]
+[assembly: AssemblyFileVersion("3.2.500")]
 
-[assembly: AssemblyInformationalVersion("3.1.505")]
+[assembly: AssemblyInformationalVersion("3.2.500")]